Fix FP16 dtype mismatch in img_fv_to_bev index assignment
With AMP, decorated_img_feat is half but output.sum() returns float.
Cast source to match destination dtype.

+++ mmdet3d/models/middle_encoders/fusion_encoder.py
@@ -1136,7 +1136,7 @@ class ISFusionEncoder(BaseModule):
         for b in range(bs):
             this_pillar_coors = pillar_coors_list[b]
             output = ref_img_pillar_cam[0][b].reshape(self.embed_dims, -1, num_points_in_pillar)
-            decorated_img_feat[b, :, this_pillar_coors[:, 2].long(), this_pillar_coors[:, 3].long()] = output.sum(dim=2)
+            decorated_img_feat[b, :, this_pillar_coors[:, 2].long(), this_pillar_coors[:, 3].long()] = output.sum(dim=2).to(decorated_img_feat.dtype)
 
         return decorated_img_feat
